test 3

+++ test/test3/test3.cpp
@@ -16,16 +16,11 @@ int main() {
         {1, 7}
     };
 
-    // Convert 2D array to array of pointers
     int *arrPtrs[3];
     for (int i = 0; i < 3; i++) {
         arrPtrs[i] = arr[i];
     }
-
-    // Sort using qsort
-    qsort(arrPtrs, 3, sizeof(int *), compare);
-
-    // Print sorted array
+    qsort(arrPtrs, 3, sizeof(double *), compare);
     for (int i = 0; i < 3; i++) {
         printf("%d %d\n", arrPtrs[i][0], arrPtrs[i][1]);
     }
